Настроил проверку email или userName

+++ src/server/routes/api.js
@@ -21,29 +21,43 @@ router.get('/load', (req, res) => {
 router.post('/registry/submit', (req, res) => {
   const userData = req.body
   userData.Date = new Date();
-  console.log(userData)
   let existingData = [];
   try {
     existingData = JSON.parse(fs.readFileSync('./bdJson/bd.json', 'utf8'));
   } catch (err) {
     console.error("Ошибка чтения файла", err);
   };
+
+  const existingMail = existingData.find(entry => entry.mail === userData.mail);
+  if (existingMail) {
+    console.log('Данный email уже используется');
+    return res.status(200).json({ error: 'mail' });
+  };
+  const existingUserName = existingData.find(entry => entry.userName === userData.userName)
+  if (existingUserName) {
+    console.log('Данный логин уже используется');
+    return res.status(200).json({ error: 'userName' });
+  };
+
   if (existingData.length !== 0) {
     userData.usrId = existingData[existingData.length - 1].usrId + 1;
   } else {
     userData.usrId = 1;
   };
+
   existingData.push(userData);
+
   console.log(existingData)
+
   fs.writeFile('./bdJson/bd.json', JSON.stringify(existingData, null, 2) , { encoding: 'utf-8'}, (err) => {
     if (err) {
       console.error('Ошибка записи файла', err);
-      res.status(500).send('Ошибка сохранения в бд');
+      res.status(500).send({ message: 'Ошибка сохранения в бд' });
     } else {
       console.log('Данные успешно сохранены в бд');
-      res.status(200).send('Успешное сохранение в бд');
+      res.status(200).send({ message: 'Успешное сохранение в бд' });
     }
-  })
-})
+  });
+});
 
 module.exports = router;
